# front_end/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "车辆维修管理系统 API",
    "version": "1.0.0",
    "description": "车辆维修管理系统的RESTful API文档",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "车辆维修管理系统API服务"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/openapi.json": {
      "get": {
        "summary": "获取OpenAPI规范文档",
        "description": "返回系统的OpenAPI 3.0规范文档JSON格式",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "成功返回OpenAPI规范文档",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "OpenAPI 3.0规范文档",
                  "properties": {
                    "openapi": {
                      "type": "string",
                      "example": "3.0.0"
                    },
                    "info": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string"
                        },
                        "version": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    },
                    "servers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "url": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "paths": {
                      "type": "object",
                      "description": "API路径定义"
                    },
                    "components": {
                      "type": "object",
                      "description": "可重用组件"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/v1/docs": {
      "get": {
        "summary": "获取Swagger UI界面",
        "description": "返回Swagger UI文档界面，用于交互式API文档浏览",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "成功返回Swagger UI页面",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "description": "Swagger UI HTML页面"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "无法加载Swagger UI"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}